refactor: move test utils for internal BPMN model in a dedicated file for reuse in other tests
